+++ packages/components/src/DateFormat/index.ts
@@ -24,9 +24,4 @@
 
  */
 
-import React, { FC } from 'react'
-import { DateTimeFormatProps, DateTimeFormat } from './DateTimeFormat'
-
-export const TimeFormat: FC<DateTimeFormatProps> = (props) => (
-  <DateTimeFormat {...props} date={false} />
-)
+export * from './DateFormat'