+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `Icon` no longer outputs a `<title />` by default and Icon is marked with `aria-hide` unless a title is explicitly specified
+- `Calendar`, `FieldDate`, `FieldDateRange`, `InputDate` & `InputDateRange` now require "deep" imports
+  - E.g.: `Calendar` imports from `@looker/components/lib/Calendar`)
 - `MenuGroup` labels improvements when using `compact`
 - `Menu` now has a new focus style for `MenuItems` as well as updateing sizing
 - `Accordion`